Made releases tar files rather than releasing to folders.  This is better for handling releases with git and also prevents problems in which git ignore causes some release files not to be commited.  Also removed toolchains relpaced with fetch script to get toolchiains from source to save git space and avoid missing .so files which were being ignored.  Scripts added for releasing toolchains and resources.

+++ setup_build_env.sh
@@ -4,35 +4,58 @@
 
 TOPDIR=`pwd`
 
+AMIROOT=`whoami | awk {'print $1'}`
+if [ "$AMIROOT" != "root" ] ; then
+
+	echo "	Error: Script must be run with root permissions "
+	exit
+fi
+
 rm -f /usr/local/alm_ref
 ln -s ${TOPDIR} /usr/local/alm_ref
 
-sudo apt-get update
+apt-get update
 
 #get cross compiler
-sudo apt-get -y install gcc-arm-linux-gnueabihf
+apt-get -y install gcc-arm-linux-gnueabihf
 
 #get u-boot tools (for mkimage to build kernel uImage)
-sudo apt-get -y install u-boot-tools
+apt-get -y install u-boot-tools
 
 #install 32 bit libs
-sudo apt-get -y install lib32z1
-sudo apt-get -y install lib32ncurses5
+apt-get -y install lib32z1
+apt-get -y install lib32ncurses5
 
 #get ncurses (needed for menuconfig)
-sudo apt-get -y install ncurses-dev
+apt-get -y install ncurses-dev
 
 #for busybox
-sudo apt-get install gcc-arm-linux-gnueabi
-apt-get install gawk
+apt-get -y install gcc-arm-linux-gnueabi
+apt-get -y install gawk
 
 #for qt
-sudo apt-get install qt-sdk
-sudo apt-get -y build-dep qt5-default
-sudo apt-get -y install build-essential perl python git
-sudo apt-get -y install "^libxcb.*" libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev
-sudo apt-get -y install libcap-dev libbz2-dev libgcrypt11-dev libpci-dev libnss3-dev build-essential libxcursor-dev libxcomposite-dev libxdamage-dev libxrandr-dev libdrm-dev libfontconfig1-dev libxtst-dev libasound2-dev gperf libcups2-dev libpulse-dev libudev-dev libssl-dev flex bison ruby
+apt-get install qt-sdk
+apt-get -y build-dep qt5-default
+apt-get -y install build-essential perl python git
+apt-get -y install "^libxcb.*" libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev
+apt-get -y install libcap-dev libbz2-dev libgcrypt11-dev libpci-dev libnss3-dev build-essential libxcursor-dev libxcomposite-dev libxdamage-dev libxrandr-dev libdrm-dev libfontconfig1-dev libxtst-dev libasound2-dev gperf libcups2-dev libpulse-dev libudev-dev libssl-dev flex bison ruby
 
 #for tslib
-sudo apt-get install autoconf
-sudo apt-get install libtool
+apt-get -y install autoconf
+apt-get -y install libtool
+
+#extract libraries
+cd ${TOPDIR}/build_tools/toolchains/
+./fetch_toolchains.sh
+
+#extract resources
+cd ${TOPDIR}/build_tools/resources/
+./extract_resources.sh
+
+cd ${TOPDIR}
+
+#init git submodules
+git submodule update --init --recursive
+
+
+
